// ui/src/generated/PresetInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PresetDevice } from "./PresetDevice";
import type { PresetMeta } from "./PresetMeta";
import type { PresetRef } from "./PresetRef";

export type PresetInfo = { preset: PresetRef, name: string, device: PresetDevice, meta: PresetMeta, };